+++ HISTORY.md
@@ -8,6 +8,10 @@
 
 - Implement Discount Model Search ({pr}`685`)
 
+#### Improvements
+
+- Fix pandas dtype in test for CategoricalArchive ({pr}`686`)
+
 ## 0.9.0
 
 To learn about this release, see our page on What's New in v0.9.0: